fix for missing Python exe
Due to the default Python exe being removed in a recent OSX release, the 'make dep' command would fail when Poetry would not be able to find the 'python' executable. This fixes the issue by creating a local virtual environment using the version of Python installed by the 'brew bundle' command.